auth/register/v2 HTTP wrapping and tests

See merge request COMP1531/21T3/groups/F10A_DODO/project-backend!67

+++ src/helper.py
@@ -51,39 +51,46 @@ def save_data_store_updates():
     Saves it to json file database
     '''
     store = data_store.get()
-    filesize = os.path.getsize("database.json")
-    # Store data_store data if json file is empty
-    if filesize == 0:
+    # Check if json database exists
+    if os.path.isfile("database.json"):
+        filesize = os.path.getsize("database.json")
+        # Store data_store data if json file is empty
+        if filesize == 0:
+            with open('database.json', 'w') as file:
+                json.dump(store, file)
+            file.close()
+        # Add new data from data_store if not empty
+        else:
+            with open("database.json") as file:
+                data = json.load(file)
+            file.close()
+            # Update 'already existing' users information 
+            for updated_user in store['users']:
+                for index, user in enumerate(data['users']):
+                    if user['u_id'] == updated_user['u_id'] and user != updated_user:
+                        data['users'][index] = updated_user
+            # Append new users
+            for user in store['users']:
+                if user not in data['users']:
+                    data['users'].append(user)
+
+            # Update 'already existing' channels information
+            for updated_chann in store['channels']:
+                for index, chann in enumerate(data['channels']):
+                    if chann['channel_id'] == updated_chann['channel_id'] and chann != updated_chann:
+                        data['channels'][index] = updated_chann
+            # Append new channels
+            for channel in store['channels']:
+                if channel not in data['channels']:
+                    data['channels'].append(channel)
+
+            with open('database.json', 'w') as file:
+                json.dump(data, file)
+            file.close()
+    # If json database does not exist create one
+    else:
         with open('database.json', 'w') as file:
             json.dump(store, file)
-        file.close()
-    # Add new data from data_store if not empty
-    else:
-        with open("database.json") as file:
-            data = json.load(file)
-        file.close()
-        # Update 'already existing' users information 
-        for updated_user in store['users']:
-            for index, user in enumerate(data['users']):
-                if user['u_id'] == updated_user['u_id'] and user != updated_user:
-                    data['users'][index] = updated_user
-        # Append new users
-        for user in store['users']:
-            if user not in data['users']:
-                data['users'].append(user)
-
-        # Update 'already existing' channels information
-        for updated_chann in store['channels']:
-            for index, chann in enumerate(data['channels']):
-                if chann['channel_id'] == updated_chann['channel_id'] and chann != updated_chann:
-                    data['channels'][index] = updated_chann
-        # Append new channels
-        for channel in store['channels']:
-            if channel not in data['channels']:
-                data['channels'].append(channel)
-
-        with open('database.json', 'w') as file:
-            json.dump(data, file)
         file.close()
 
 def save_database_updates(updated_database):
@@ -103,12 +110,46 @@ def get_data():
     Returns:
         dictionary in a smilar manner to datastore.get
     '''
-    filesize = os.path.getsize("database.json")
-    if filesize == 0:
-        return {}
-    with open('database.json') as file:
-        data = json.load(file)
-    file.close()
-    return data
+    # Check if database exists
+    if os.path.isfile("database.json"):
+        filesize = os.path.getsize("database.json")
+        if filesize == 0:
+            return {}
+        with open('database.json') as file:
+            data = json.load(file)
+        file.close()
+        return data
+    else:
+        return{}
 
 
+def create_handle(name_first, name_last, data):
+
+    '''
+    Generate handle
+    Arguments:
+        name_first (string)
+        name_last (string)
+        data (dictionary) - database
+    Return:
+        handle_str (string)
+    '''
+    handle_str = name_first.lower() + name_last.lower()
+    handle_str = ''.join(char for char in handle_str if char.isalnum())
+    handle_rep_num = -1
+    if len(handle_str) > 20:
+        handle_str = handle_str[0:20]
+    if len(data['users']) != 0:
+        for user in data['users']:
+            if user['handle_str'][0:len(handle_str)] == handle_str:
+                handle_rep_num += 1
+        if handle_rep_num != -1:
+            handle_str = handle_str + str(handle_rep_num)
+    return handle_str
+
+def create_permission_id(data):
+    if len(data['users']) == 0:
+        permission_id = 1
+    else:
+        permission_id = 2
+    return permission_id
